% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtEff.R
\name{DEFF}
\alias{DEFF}
\title{Calculate DEFF}
\usage{
DEFF(x)
}
\arguments{
\item{x}{= weights vector (name of weights column)}
}
\value{
Design effect (DEFF)
}
\description{
Calculates design effect (DEFF)
}
\examples{
DEFF(testweights$weights_column)
}
\references{
Design effect (DEFF) due to weighting => n * (sum(x^2) / sum(x)^2)
}
