% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtEff.R
\name{DEFT}
\alias{DEFT}
\title{Calculate DEFT}
\usage{
DEFT(x)
}
\arguments{
\item{x}{= weights vector (name of weights column)}
}
\value{
Root design effect (DEFT)
}
\description{
Calculates root design effect (DEFT)
}
\examples{
DEFT(testweights$weights_column)
}
\references{
Root design effect (DEFT) => square root of DEFF
}
