% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtEff.R
\name{ESS}
\alias{ESS}
\title{Calculate ESS}
\usage{
ESS(x)
}
\arguments{
\item{x}{= weights vector (name of weights column)}
}
\value{
Effective sample size (ESS)
}
\description{
Calculates effective sample size (ESS)
}
\examples{
ESS(testweights$weights_column)
}
\references{
Effective sample size (ESS) => sum(x)^2 / sum(x^2)
}
