% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtEff.R
\name{FULL}
\alias{FULL}
\title{Calculate Full Statistics}
\usage{
FULL(p = 50, conf = 95, N, wtcol)
}
\arguments{
\item{p}{= percentage for which MOE is calculated (optional, default is p = 50)}

\item{conf}{= level of confidence (optional, default is conf = 95)}

\item{N}{= population size (optional, used for finite population correction)}

\item{wtcol}{= Weights vector (name of weights column)}
}
\value{
DEFF, weighting loss, DEFT, ESS, and MOE
}
\description{
Calculates DEFF, weighting loss, DEFT, ESS, and MOE
}
\examples{
FULL(N=3000, wtcol=testweights$weights_column)
}
