% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtEff.R
\name{MOE}
\alias{MOE}
\title{Calculate MOE}
\usage{
MOE(p = 50, conf = 95, N, wtcol)
}
\arguments{
\item{p}{= percentage for which MOE is calculated (optional, default is p = 50)}

\item{conf}{= level of confidence (optional, default is conf = 95)}

\item{N}{= population size (optional, used for finite population correction)}

\item{wtcol}{= Weights vector (name of weights column)}
}
\value{
Weighted margin of error (MOE)
}
\description{
Calculates weighted margin of error (MOE)
}
\examples{
MOE(N=3000, wtcol=testweights$weights_column)
}
\references{
Weighted margin of error (MOE) => unweighted MOE * DEFT
}
