% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WgtEff.R
\name{WTGLOSS}
\alias{WTGLOSS}
\title{Calculate weighting loss}
\usage{
WTGLOSS(x)
}
\arguments{
\item{x}{= weights vector (name of weights column)}
}
\value{
Weighting loss
}
\description{
Calculates weighting loss
}
\examples{
WTGLOSS(testweights$weights_column)
}
\references{
Weighting loss => DEFF-1
}
