% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testweights.R
\docType{data}
\name{testweights}
\alias{testweights}
\title{An example weights column for a data set of 320 cases}
\format{A data frame with 320 rows and 1 variable
\describe{\item{weights_column}{data weights}
}}
\source{
Example data generated by author
}
\usage{
testweights
}
\description{
An example weights column for a data set of 320 cases
}
\keyword{datasets}
